+++ src/connected/Viewer.tsx
@@ -7,6 +7,7 @@
 import { connect } from "react-redux";
 import { default as OlMap } from "ol/Map";
 
+import { Config } from "@/config";
 import { AppState } from "@/states/appState";
 import {
   baseMapLayersSelector,
@@ -35,6 +36,7 @@ import ColorBarLegend2 from "./ColorBarLegend2";
 import MapSplitter from "./MapSplitter";
 import MapPointInfoBox from "./MapPointInfoBox";
 import MapControlActions from "./MapControlActions";
+import ZoomBox from "./ZoomBox";
 
 interface OwnProps {
   onMapRef?: (map: OlMap | null) => void;
@@ -68,6 +70,8 @@ const mapStateToProps = (state: AppState, ownProps: OwnProps) => {
     imageSmoothing: imageSmoothingSelector(state),
     variableSplitPos: state.controlState.variableSplitPos,
     onMapRef: ownProps.onMapRef,
+    allowZoomBox: Config.instance.branding.allowZoomBox,
+    zoomBox: <ZoomBox />,
   };
 };